when the when you when when

+++ shaders/bufferSettings.glsl
@@ -27,7 +27,7 @@ const int colortex4Format         = RGBA32F;
 const bool colortex4Clear         = false;
 
 MAIN COLOR BUFFER
-const int colortex5Format         = RGBA32F;
+const int colortex5Format         = RGBA16F;
 const bool colortex5MipmapEnabled = true;
 
 VOLUMETRIC LIGHTING / SKY ILLUMINANCE
